requirements.txt
duckduckgo-search>=6.0
feedparser>=6.0
beautifulsoup4>=4.12
requests>=2.31
numpy>=1.26
pandas>=2.2
streamlit>=1.40
plotly>=5.24
python-dotenv>=1.0
lxml>=5.0
google-genai>=1.0
langchain-openai>=0.3
